use C standard integer types

+++ iodevices/altair-88-2sio.c
@@ -25,6 +25,7 @@
  * 19-JUL-2020 avoid problems with some third party terminal emulations
  */
 
+#include <stdint.h>
 #include <unistd.h>
 #include <stdio.h>
 #include <ctype.h>
@@ -41,7 +42,7 @@
 
 #define BAUDTIME 10000000
 
-extern unsigned long long get_clock_us(void);
+extern uint64_t get_clock_us(void);
 
 static const char *TAG = "2SIO";
 
@@ -50,7 +51,7 @@ int sio1_strip_parity;
 int sio1_drop_nulls;
 int sio1_baud_rate = 115200;
 
-static unsigned long long sio1_t1, sio1_t2;
+static uint64_t sio1_t1, sio1_t2;
 static BYTE sio1_stat;
 
 int sio2_upper_case;
@@ -58,7 +59,7 @@ int sio2_strip_parity;
 int sio2_drop_nulls;
 int sio2_baud_rate = 115200;
 
-static unsigned long long sio2_t1, sio2_t2;
+static uint64_t sio2_t1, sio2_t2;
 static BYTE sio2_stat;
 
 /*
