+++ z80core/simz80-dd.c
@@ -10,6 +10,7 @@
  *	starting with 0xdd
  */
 
+#include <stdint.h>
 #include "sim.h"
 #include "simglb.h"
 #include "config.h"
@@ -337,8 +338,8 @@ int op_dd_handle(void)
 
 	register int t;
 #ifdef FRONTPANEL
-	extern unsigned long long get_clock_us(void);
-	unsigned long long clk;
+	extern uint64_t get_clock_us(void);
+	uint64_t clk;
 #endif
 
 #ifdef BUS_8080
